Changed when updateTouchPoint() runs
updateTouchPoint() now runs only when the user is not currently
selecting a new point and only after corners have been found.

Also commented out morphological closing operation for now, since it
causes strange stuff to happen when the subject blob nears an edge of
the frame.

And added in the updated intersection point formula, along with a bunch
of test code for the Hough lines method (that isn't totally working yet)